requirements.txt
pillow
openai
apscheduler
requests